Use new input everywhere, allow for clear button, disabled, hitCopy and aligned status messages
