+++ common/views/components/SearchBar/SearchBar.tsx
@@ -3,7 +3,6 @@ import {
   FunctionComponent,
   RefObject,
   SetStateAction,
-  useContext,
   useRef,
 } from 'react';
 import styled from 'styled-components';
@@ -12,12 +11,10 @@ import ButtonSolid, {
   ButtonTypes,
 } from '@weco/common/views/components/ButtonSolid/ButtonSolid';
 import { themeValues } from '@weco/common/views/themes/config';
-import ClearSearch from '@weco/common/views/components/ClearSearch/ClearSearch';
-import { AppContext } from '@weco/common/views/components/AppContext/AppContext';
 
 const Container = styled.div`
   display: flex;
-  align-items: stretch;
+  align-items: flex-end;
 `;
 const SearchInputWrapper = styled.div`
   flex: 1 1 auto;
@@ -56,7 +53,6 @@ const SearchBar: FunctionComponent<Props> = ({
   inputRef,
   location,
 }) => {
-  const { isEnhanced } = useContext(AppContext);
   const defaultInputRef = useRef<HTMLInputElement>(null);
 
   return (
@@ -71,15 +67,9 @@ const SearchBar: FunctionComponent<Props> = ({
           setValue={setInputValue}
           ref={inputRef || defaultInputRef}
           form={form}
-          big={true}
+          big
+          hasClearButton
         />
-        {inputValue && isEnhanced && (
-          <ClearSearch
-            inputRef={inputRef || defaultInputRef}
-            setValue={setInputValue}
-            right={16}
-          />
-        )}
       </SearchInputWrapper>
       <SearchButtonWrapper>
         <ButtonSolid
