+++ common/views/components/TextInput/index.tsx
@@ -4,84 +4,71 @@ import {
   ForwardRefRenderFunction,
   ChangeEvent,
   FocusEvent,
+  useContext,
 } from 'react';
 import styled from 'styled-components';
 import Icon from '@weco/common/views/components/Icon/Icon';
-import { check } from '@weco/common/icons';
+import { exclamation, tickCircle } from '@weco/common/icons';
 import { font } from '@weco/common/utils/classnames';
+import ClearInput from '@weco/common/views/components/TextInput/TextInput.Clear';
+import Space from '../styled/Space';
+import { AppContext } from '../AppContext/AppContext';
+
+export const TextInputLabel = styled.label.attrs({
+  className: font('intb', 5),
+})`
+  display: block;
+  line-height: 1.2;
+`;
 
 type TextInputWrapProps = {
+  $status?: 'error' | 'success';
   $big?: boolean;
-  $hasErrorBorder: boolean;
-  $darkBg?: boolean;
+  $isDisabled?: boolean;
 };
-export const TextInputWrap = styled.div.attrs<TextInputWrapProps>(props => ({
+export const TextInputWrap = styled(Space).attrs<TextInputWrapProps>(props => ({
   className: props.$big ? font('intr', 4) : font('intr', 5),
+  $v: { size: 's', properties: ['margin-top'] },
 }))<TextInputWrapProps>`
   display: flex;
   position: relative;
-  border: ${props =>
-    props.$hasErrorBorder
-      ? `3px solid ${props.theme.color('validation.red')}`
-      : `2px solid ${props.theme.color(
-          props.$darkBg ? 'white' : 'neutral.600'
-        )}`};
+  border-width: 1px;
+  border-style: solid;
+  border-color: ${props =>
+    props.$status
+      ? props.$status === 'error'
+        ? props.theme.color('validation.red')
+        : props.theme.color('validation.green')
+      : props.theme.color('neutral.600')};
 
   &:focus-within {
-    box-shadow: ${props => props.theme.focusBoxShadow};
+    box-shadow: 0 0 0 6px ${props => props.theme.color('focus.yellow')};
+    outline: 3px solid ${props => props.theme.color('black')};
   }
 
   overflow: hidden;
-`;
 
-type TextInputLabelProps = {
-  $isEnhanced: boolean;
-  $hasValue: boolean;
-};
-export const TextInputLabel = styled.label<TextInputLabelProps>`
-  position: absolute;
-  left: 15px;
-
-  /* Styles for browsers that don't support :focus-within (<=IE11) */
-  font-size: 14px;
-  transform: translateY(0%);
-  top: 4px;
-
-  /* IE doesn't support :focus-within, but you can't test for :focus-within
-  using @supports. Fortunately, IE doesn't support @supports, so this only
-  targets browsers that support @supports (> IE11) */
-  @supports (display: block) {
-    font-size: inherit;
-    transform: translateY(-50%);
-    top: 50%;
+  &:hover {
+    border-color: ${props => props.theme.color('black')};
   }
 
-  color: ${props => props.theme.color('neutral.600')};
-  white-space: nowrap;
-  transition:
-    top ${props => props.theme.transitionProperties},
-    font-size ${props => props.theme.transitionProperties},
-    transform ${props => props.theme.transitionProperties};
-  pointer-events: none;
-
   ${props =>
-    (!props.$isEnhanced || props.$hasValue) &&
-    `
-    @supports (display: block) {
-      top: 4px;
-      transform: translateY(0);
-      font-size: 14px;
-    }
-  `}
+    props.$isDisabled
+      ? `border-color:  ${props.theme.color('neutral.500')}`
+      : ``}
 `;
 
-type TextInputInputProps = {
-  $hasErrorBorder: boolean;
-};
+const HintCopy = styled.span.attrs({
+  className: font('intr', 5),
+})`
+  display: block;
+  color: ${props => props.theme.color('neutral.700')};
+`;
+
 export const TextInputInput = styled.input.attrs<{ $type?: string }>(props => ({
   type: props.$type || 'text',
-}))<TextInputInputProps>`
-  padding: 17px 35px 17px 15px;
+}))`
+  padding: 10px 36px 10px 16px;
   appearance: none;
   border: 0;
   height: 100%;
@@ -91,31 +78,40 @@ export const TextInputInput = styled.input.attrs<{ $type?: string }>(props => ({
   &::-ms-clear {
     display: none;
   }
+
+  &:disabled {
+    background-color: ${props => props.theme.color('neutral.300')};
+  }
 `;
 
-const TextInputCheckmark = styled.span.attrs<{
-  'data-testid'?: 'TextInputCheckmark';
-}>({
-  'data-testid': 'TextInputCheckmark',
+const StatusMessage = styled(Space).attrs({
+  className: font('intr', 6),
+  $v: { size: 's', properties: ['margin-top'] },
 })`
-  position: absolute;
-  top: 50%;
-  transform: translateY(-50%);
-  right: 10px;
-  line-height: 0;
+  display: flex;
+  align-items: center;
+
+  p {
+    display: inline-block;
+    ${props => props.theme.makeSpacePropertyValues('s', ['margin-left'])};
+    margin-bottom: 0;
+  }
 `;
 
-export const TextInputErrorMessage = styled.span.attrs({
-  role: 'alert',
-  'data-testid': 'TextInputErrorMessage',
-  className: 'font-intb',
-})`
-  display: block;
-  font-size: 14px;
-  margin-top: 10px;
-  padding-left: 15px;
-  color: ${props => props.theme.color('validation.red')};
-`;
+export const InputErrorMessage = ({
+  id,
+  errorMessage,
+}: {
+  id?: string;
+  errorMessage: string;
+}) => {
+  return (
+    <StatusMessage data-testid="TextInputErrorMessage" id={id} role="alert">
+      <Icon icon={exclamation} iconColor="validation.red" />
+      <p>{errorMessage}</p>
+    </StatusMessage>
+  );
+};
 
 type Props = {
   label: string;
@@ -126,8 +122,11 @@ type Props = {
   type?: string;
   pattern?: string;
   required?: boolean;
+  disabled?: boolean;
   placeholder?: string;
+  hintCopy?: string;
   errorMessage?: string;
+  successMessage?: string;
   isValid?: boolean;
   setIsValid?: (value: boolean) => void;
   showValidity?: boolean;
@@ -137,7 +136,8 @@ type Props = {
   ariaLabel?: string;
   ariaDescribedBy?: string;
   form?: string;
-  darkBg?: boolean;
+  hasClearButton?: boolean;
+  clearHandler?: () => void;
 };
 
 const Input: ForwardRefRenderFunction<HTMLInputElement, Props> = (
@@ -150,8 +150,11 @@ const Input: ForwardRefRenderFunction<HTMLInputElement, Props> = (
     name,
     pattern,
     required,
+    disabled,
     placeholder,
+    hintCopy,
     errorMessage,
+    successMessage,
     isValid,
     setIsValid,
     showValidity,
@@ -161,10 +164,13 @@ const Input: ForwardRefRenderFunction<HTMLInputElement, Props> = (
     ariaDescribedBy,
     form,
     big,
-    darkBg,
+    hasClearButton,
+    clearHandler,
   }: Props,
   ref: RefObject<HTMLInputElement>
 ) => {
+  const { isEnhanced } = useContext(AppContext);
+
   function onChange(event: ChangeEvent<HTMLInputElement>) {
     const isValueValid = event.currentTarget.validity.valid;
 
@@ -188,43 +194,68 @@ const Input: ForwardRefRenderFunction<HTMLInputElement, Props> = (
 
   return (
     <div>
+      <TextInputLabel htmlFor={id}>{label}</TextInputLabel>
+
+      {hintCopy && (
+        <Space $v={{ size: 's', properties: ['margin-top'] }}>
+          <HintCopy>{hintCopy}</HintCopy>
+        </Space>
+      )}
+
       <TextInputWrap
-        $hasErrorBorder={!!(!isValid && showValidity)}
         $big={!!big}
-        $darkBg={darkBg}
+        $isDisabled={disabled}
+        $status={
+          !isValid && showValidity
+            ? 'error'
+            : isValid && showValidity
+            ? 'success'
+            : undefined
+        }
       >
-        <label className="visually-hidden" htmlFor={id}>
-          {label}
-        </label>
         <TextInputInput
           ref={ref}
           id={id}
           name={name}
           required={required}
+          disabled={disabled}
           value={value}
           pattern={pattern}
           onChange={onChange}
           onBlur={onBlur}
-          placeholder={placeholder || label}
+          placeholder={placeholder}
           autoFocus={autoFocus}
           $type={type}
-          $hasErrorBorder={!!(!isValid && showValidity)}
           aria-label={ariaLabel}
           aria-describedby={ariaDescribedBy}
           aria-invalid={!!(!isValid && showValidity)}
           aria-errormessage={errorMessage && `${id}-errormessage`}
           form={form}
         />
-        {isValid && showValidity && (
-          <TextInputCheckmark>
-            <Icon icon={check} iconColor="validation.green" />
-          </TextInputCheckmark>
+        {isEnhanced && hasClearButton && value !== '' && (
+          <ClearInput
+            inputRef={ref}
+            clickHandler={clearHandler}
+            setValue={setValue}
+            right={10}
+          />
         )}
       </TextInputWrap>
+      {successMessage && isValid && showValidity && (
+        <StatusMessage>
+          <Icon
+            data-testid="TextInputCheckmark"
+            icon={tickCircle}
+            iconColor="validation.green"
+          />
+          <p>{successMessage}</p>
+        </StatusMessage>
+      )}
       {errorMessage && !isValid && showValidity && (
-        <TextInputErrorMessage id={`${id}-errormessage`} role="alert">
-          {errorMessage}
-        </TextInputErrorMessage>
+        <InputErrorMessage
+          id={`${id}-errormessage`}
+          errorMessage={errorMessage}
+        />
       )}
     </div>
   );
