+++ identity/webapp/src/frontend/MyAccount/ChangeEmail.tsx
@@ -3,7 +3,7 @@ import { useForm, Controller } from 'react-hook-form';
 import { ErrorMessage } from '@hookform/error-message';
 import { FieldMargin } from '../components/Form.style';
 import TextInput, {
-  TextInputErrorMessage,
+  InputErrorMessage,
 } from '@weco/common/views/components/TextInput';
 import { PasswordInput } from '../components/PasswordInput';
 import { validEmailPattern } from '../components/ValidationPatterns';
@@ -150,7 +150,7 @@ export const ChangeEmail: React.FunctionComponent<
             errors={formState.errors}
             name="password"
             render={({ message }) => (
-              <TextInputErrorMessage>{message}</TextInputErrorMessage>
+              <InputErrorMessage errorMessage={message} />
             )}
           />
         </FieldMargin>
