+++ identity/webapp/src/frontend/MyAccount/ChangePassword.tsx
@@ -6,7 +6,7 @@ import { FieldMargin } from '../components/Form.style';
 import ButtonSolid, {
   ButtonTypes,
 } from '@weco/common/views/components/ButtonSolid/ButtonSolid';
-import { TextInputErrorMessage } from '@weco/common/views/components/TextInput';
+import { InputErrorMessage } from '@weco/common/views/components/TextInput';
 import { useForm, useWatch } from 'react-hook-form';
 import { ModalContainer, ModalTitle, StatusAlert } from './MyAccount.style';
 import { ChangeDetailsModalContentProps } from './ChangeDetailsModal';
@@ -134,7 +134,7 @@ export const ChangePassword: React.FunctionComponent<
             errors={formState.errors}
             name="password"
             render={({ message }) => (
-              <TextInputErrorMessage>{message}</TextInputErrorMessage>
+              <InputErrorMessage errorMessage={message} />
             )}
           />
         </FieldMargin>
@@ -156,7 +156,7 @@ export const ChangePassword: React.FunctionComponent<
             errors={formState.errors}
             name="newPassword"
             render={({ message }) => (
-              <TextInputErrorMessage>{message}</TextInputErrorMessage>
+              <InputErrorMessage errorMessage={message} />
             )}
           />
         </FieldMargin>
@@ -184,7 +184,7 @@ export const ChangePassword: React.FunctionComponent<
             errors={formState.errors}
             name="confirmation"
             render={({ message }) => (
-              <TextInputErrorMessage>{message}</TextInputErrorMessage>
+              <InputErrorMessage errorMessage={message} />
             )}
           />
           <Space $v={{ size: 's', properties: ['margin-top'] }}>
